Extract widgets into file
